Feat(issue): Update find command to search by tags and description as well
