+++ src/main/java/seedu/address/logic/parser/issue/FindIssueCommandParser.java
@@ -7,7 +7,7 @@ import java.util.Arrays;
 import seedu.address.logic.commands.issue.FindIssueCommand;
 import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.issue.RoomNumberContainsKeywordsPredicate;
+import seedu.address.model.issue.RoomNumberOrTagContainsKeywordsPredicate;
 
 /**
  * Parses input arguments and creates a new FindIssueCommand object
@@ -29,7 +29,7 @@ public class FindIssueCommandParser implements Parser<FindIssueCommand> {
 
         String[] nameKeywords = trimmedArgs.split("\\s+");
 
-        return new FindIssueCommand(new RoomNumberContainsKeywordsPredicate(Arrays.asList(nameKeywords)));
+        return new FindIssueCommand(new RoomNumberOrTagContainsKeywordsPredicate(Arrays.asList(nameKeywords)));
     }
 
 }
